# src/main/java/com/salygin/ssubench/api/error/ApiErrorResponse.java
package com.salygin.ssubench.api.error;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_NULL)
public record ApiErrorResponse(
        int status,
        String error,
        String message,
        String path,
        String trackingId
) {}